Test: backend specification on operations test (#25)
Right now as SSE is supported on almost all the CPUs testing it along with the simple backend gave no issues. But with the introduction of ARM SIMD or AVX512 not all CPUs support it. So when testing, the selected backend should be passed to test only that, and not all the implemented ones. As commented on issue #22.